[sgen] Finish concurrent M&S quicker
When we finish the concurrent marking, we leave the mutator to trigger a nursery collection, where we can see that the worker thread is done and finish the major collection. We speed up this process by invalidating the nursery fragments which leads to a collection happening at the next nursery memory request, compared to happening at nursery fill time. We do this in order to avoid further allocation and card table marking which would increase the finishing pause time.

We might be able to trigger a collection from the worker thread, but we try to avoid this for now since it could lead to problems. Especially if we were to add multiple worker threads later on.