[sgen] Preclean mod union entries before the finishing pause
When the concurrent mark ends, we have another concurrent phase where we scan the cardtable trying to clean as many cards as possible. Scanning the objects (or valuetypes) in a card can lead to dirtying not only of the same card but also of a previous or next card (if the object is contained in multiple cards, since we scan the entire object). Since we only scan in increasing memory addresses, we only need to worry about not cleaning next cards marked in the precleaning phase (since the objects that lead to the marking won't be rescanned and we would lose the remset). We handle this by always recording the highest card that was marked and preclean only if it is smaller than the current card.

Due to the way we are scanning objects that are bigger than a card, we also need to make sure we don't preclean the last card of the object, since that card might be marked due to a remset in the next object. Cleaning it would stop us from scanning the next object. The precleaning of the card will happen when scanning the next object (provided that the scanning of the first object didn't lead to marking of that card, which we would detect as described above)