[utils] Fix freeing of descriptor with used slots for lock free allocator
We allocate on a block by acquiring the descriptor for it (either cas-ing on the active descriptor of the allocator or popping from the partial list). When a thread acquires a descriptor, that thread is the only one allocating in the block but other threads can still free inside the block.

When we were freeing inside the block we checked if the block got empty afterwards, acquired it in that case and freed it. We need to check whether, after acquiring the descriptor for the block, it is still empty, because before acquiring the desired descriptor, another thread might have cleaned it and reallocated it at the same addresss, leading us to incorrectly try to free it, even though it is a logically different descriptor.