[runtime] Scan MonoGHashTable roots generationally
We allocate the vector that holds the key/value pairs in managed memory. When scanning the root we only ask to scan this array by the GC, which, in the case of nursery collections, will not do anything since most likely the array is in the old gen. The actual entries are going to be scanned in the card table scanning process. (This means that each insertion/rehashing in the hash table will lead to dirtying of cards)