[runtime] Fix detecting abort at end of abort protected block
If we have a pending abort then we know we will have the thread_state set and the thread_interruption_requested bumped. We can then just use mono_thread_execute_interruption to start processing the exception. mono_thread_resume_interruption is not fit for the job since it attempts to do interruption setting (which is already done) and bailing out afterwards.