[sgen] Promotion bailout
This is buggy WIP.

It seems that the amount of copied objects in a collection are a pretty good indicator regarding the collection duration. This means that, during an actual collection, we can predict a worst case scenario nursery time = current_collection_time * nursery_size / current_promoted_size, and give up on promotion if we consider that it might take too long.

Since we no longer promote objects, we need to pin them instead, in order to mark them as visited. The problem is that we would need to also unpin the vtable before resuming the world. We can do this either by fully scanning the nursery (which sounds unacceptable) or having them stored somehow, so we can traverse them later (or having some separate marking option/gc algorithm, which sounds super complicated). Having the refs stored in a data structure is unaccetable since the array can get super large and it turns out to actually double the collection time (instead of reducing it). The only reasonable alternative would be to use the sync field of the object to link together objects that are to be unpinned and for the few objects that have the field occupied, we can just enqueue the refs to the pin queue (which would also need to be made parallel).

For now, implementing this sounds like too much complication considering that we wouldn't use this mode anyway by default (since it can lead to mutator slowing down quite a bit), and for low pause scenarios we have the trivial alternative of just resizing the nursery more aggresively in order to favor low pause times.