[coop] Fix safehandle crashes
This fixes the safehandle crashes for me. I haven't been able to reproduce this with the necessary logging enabled, but I can't say I really tried yet. So I don't have a decisive conclusion, but the crashes only seem to happen in coop mode. The fact that this simple commit also fixes the crashes leads me to believe that the GC fails to pin some objects because coop doesn't properly initialize the required context, used by the GC to scan stack and regs.